+++ package.json
@@ -13,8 +13,6 @@
     "@emotion/react": "^11.13.0",
     "@emotion/styled": "^11.13.0",
     "@chakra-ui/react": "^2.8.2",
-    "@emotion/react": "^11.13.0",
-    "@emotion/styled": "^11.13.0",
     "framer-motion": "^11.3.28",
     "axios": "^1.7.4",
     "react": "^18.3.1",
