+++ src/components/loginUser/loginForm.module.css
@@ -1,4 +1,3 @@
-/* loginForm.module.css */
 .body {
     display: flex;
     justify-content: center;
