+++ src/components/registerUser/registerForm.module.css
@@ -1,4 +1,3 @@
-/* registerForm.module.css */
 .body {
     display: flex;
     justify-content: center;
